systems: move system state classes into system files
This change moves the location of some component implementations from
the entities/ directory to the systems/ directory. For some systems,
a single per-entity component contains most of the relevant state, and
this change moves the implementations of these components so they're
next to the systems that use them.

In general, this continues the philosophy that entities are nothing
special: components and systems should contain the bulk of simulation
state and logic, and files dedicated to entities should resemble
factories.
