+++ src/entities/pickups/index.ts
@@ -1,5 +1,5 @@
-import { PickupType } from '~/entities/pickup'
 import { makeNadaPickup } from '~/entities/pickups/Nada'
+import { PickupType } from '~/systems/pickups'
 
 export const PickupConstructors = {
   [PickupType.Core]: makeNadaPickup,
