--- src/entities/turret/Turret.ts
@@ -1,7 +1,0 @@
-export class Turret {
-  cooldownTtl: number
-
-  constructor() {
-    this.cooldownTtl = 0
-  }
-}
